[Http] Moving RequestHandler classes to a new Server\Handler namespace for better organization.
