+++ tests/Functional/ApplicationTest.php
@@ -19,7 +19,7 @@ use Valkyrja\Filesystem\Manager\Contract\Filesystem;
 use Valkyrja\Http\Client\Manager\Contract\Client;
 use Valkyrja\Http\Message\Factory\Contract\ResponseFactory;
 use Valkyrja\Http\Routing\Dispatcher\Contract\Router;
-use Valkyrja\Http\Server\Contract\RequestHandler;
+use Valkyrja\Http\Server\Handler\Contract\RequestHandler;
 use Valkyrja\Log\Logger\Contract\Logger;
 use Valkyrja\Session\Manager\Contract\Session;
 use Valkyrja\Tests\EnvClass;
